+++ src/Components/Testimonial.jsx
@@ -1,0 +1,3 @@
+export default function Testimonial() {
+  return <section className="testimonial section"></section>;
+}
