+++ src/Components/Work.jsx
@@ -1,0 +1,3 @@
+export default function Work() {
+  return <section className="work section"></section>;
+}
